dir: move to domain
