+++ src/main/java/lotto/domain/entity/Lotto.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain.entity;
 
 import java.util.List;
